requirements.txt
fastapi
uvicorn
langchain
openai
youtube-transcript-api
faiss-cpu
python-dotenv
